pyproject.toml
[project]
name = "smu-auto-evaluation"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "ddddocr>=1.5.6",
    "pillow>=11.3.0",
    "pycryptodome>=3.23.0",
    "requests>=2.32.5",
]
